requirements.txt
numpy
imageio
scipy
pytorch
torchvision
pillow